requirements.txt
fastapi
fastapi[standard]
sqlalchemy
alembic
pydantic-settings
pwdlib[argon2]
python-multipart
pyjwt